Remove setlocale call in __hxcpp_stdlibs_boot (#1284)
On Windows, this meant the utf-8 string passed to printf("%s") was
interpreted as if it is from the default system locale encoding instead
of utf-8, which are incompatible so any non-ASCII text comes out
garbled. This setlocale call also meant that using `chcp 65001` does not
haxe the correct effect of being able to view the full range of utf-8
characters from print output.

Originally, this setlocale call was required required due to the use of
print("%S"), but the printing functions in Stdlibs.cpp no longer rely on
that so it is safe to remove.